+++ swrlib/Data/Item.cs
@@ -4,9 +4,9 @@ public class Item
 {
     public int Month { get; set; }
     public int Year { get; set; }
-    public double Value { get; set; }
+    public float Value { get; set; }
 
-    public Item(int month, int year, double value)
+    public Item(int month, int year, float value)
     {
         Month = month;
         Year = year;
